Unemployment and depression

A subset of data from Ginexi and colleagues (2000) measuring changes in depressive symptoms after job loss in a sample of 254 recently unemployed men and women. Interviews were conducted in three waves at around 1, 5, and 12 months after job loss.

Usage

depression_unemployment

Format

A person-period data frame with 674 rows and 5 columns:

id

Participant ID.

interview

Time of interview.

months

Months since job loss.

depression

Center for Epidemiologic Studies' Depression (CES-D) scale score (Radloff, 1977)

unemployed

Binary indicator for whether the participant was unemployed at time of interview.

Source

Ginexi, E. M., Howe, G. W., & Caplan, R. D. (2000). Depression and control beliefs in relation to reemployment: What are the directions of effect? Journal of Occupational Health Psychology, 5, 323–336. https://doi.org/10.1037//1076-8998.5.3.323

References

Radloff, L. S. (1977). The CES-D scale: A self report major depressive disorder scale for research in the general population. Applied Psychological Measurement, 1, 385–401. https://doi.org/10.1177/014662167700100306