Age of first sexual intercourse

A subset of data from Capaldi, Crosby, and Stoolmiller (1996) measuring the grade year of first sexual intercourse in a sample of 180 at-risk heterosexual adolescent males. Adolescent males were followed from Grade 7 up to Grade 12 or until they reported having had sexual intercourse for the first time.

Usage

first_sex

Format

A person-level data frame with 180 rows and 5 columns:

id

Participant ID.

grade

Grade year of first sexual intercourse.

censor

Censoring status.

parental_transition

Binary indicator for whether the adolescent experienced a parental transition (where their parents separated or repartnered).

parental_antisociality

Composite score across four indicators measuring parents' level of antisocial behaviour during the child's formative years.

Source

Capaldi, D. M., Crosby, L., & Stoolmiller, M. (1996). Predicting the timing of first sexual intercourse for at-risk adolescent males. Child Development, 67, 344–359. https://doi.org/10.2307/1131818

Note

Capaldi, Crosby, and Stoolmiller's (1996) original sample consisted of 182 adolescent males after applying exclusion criteria for analysis; Singer and Willett (2003) excluded an additional two males from the data who reported having anal intercourse with another male.