Time to horn honking

A subset of data from Diekmann and colleagues (1996) measuring time to horn honking in a sample of 57 motorists who were purposefully blocked at a green light by a Volkswagen Jetta at a busy intersection near the centre of Munich, West Germany on two busy afternoons (Sunday and Monday) in 1998. Motorists were followed until they honked their horns or took an alternative action (beaming or changing lanes).

Usage

honking

Format

A person-level data frame with 57 rows and 3 columns:

id

Participant ID.

seconds

Number of seconds to horn honking or an alternative action.

censor

Censoring status.

Source

Diekmann, A., Jungbauer-Gans, M., Krassnig, H., & Lorenz, S. (1996). Social status and aggression: A field study analyzed by survival analysis. Journal of Social Psychology, 136, 761–768. https://doi.org/10.1080/00224545.1996.9712252