Supreme Court justice tenure

Data from Zorn and Van Winkle (2000) how long all 107 justices appointed to the U.S. Supreme Court between 1789 and 1980 remained in their position.

Usage

judges

Format

A person-level data frame with 109 rows and 7 columns:

id

Justice ID.

tenure

Time to retirement or death in years.

dead

Binary indicator for whether the justice died.

retired

Binary indicator for whether the justice retired.

left_appointment

Binary indicator for whether the justice left their appointment.

appointment_age

Age at time of appointment.

appointment_year

Year of appointment.

Source

Zorn, C. J., & van Winkle, S. R. (2000). A competing risks model of Supreme Court vacancies, 1780–1992. Political Behavior, 22, 145–166.