Age of first suicide ideation

A subset of data from Bolger and colleagues (1989) measuring age of first suicide ideation in a sample of 391 undergraduate students aged 16 to 22. Age of first suicide ideation was measured with a two-item survey asking respondents "Have you ever thought of committing suicide?" and if so, "At what age did the thought first occur to you?"

Usage

suicide_ideation

Format

A person-level data frame with 391 rows and 4 columns:

id

Participant ID.

age

Reported age of first suicide ideation.

censor

Censoring status.

age_now

Participant age at the time of the survey.

Source

Bolger, N., Downey, G., Walker, E., & Steininger, P. (1989). The onset of suicide ideation in childhood and adolescence. Journal of Youth and Adolescence, 18, 175–189. https://doi.org/10.1007/BF02138799